[stable/concourse] optionally uses deployment instead of statefulset
When making complete use of the ephemerality capability of
concourse workers, it's better to not have a name that gets
reused everytime a worker comes up (behavior that statefulsets
give us), meaning that a Deployment makes more sense in such
case.

The commit introduces a `-deployment` object for the scenario
where persistence for the workers is not desired.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>