[stable/concourse] Makes use of the worker healthcheck endpoint
Previously, the liveness probe of the worker was based on logs, which
could end up killing the whole worker in the case of a malformed
pipeline.

Now, making use of the native concourse healthchecking that the workers
provide, we can delegate to concourse the task of telling k8s if it's
alive or not.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Signed-off-by: Saman Alvi <salvi@pivotal.io>