[stable/concourse] Removes GARDEN_ env in favor of config file
On the 5.x series of Concourse, there's no need to have all the garden
flags specified as environment variables anymore.

That's because the new image assumes that a `gdn` binary is shipped
together, which can look for a set of configurations from a specific
config file.

The set of possible values that can be used in the configuration file
can be found here [1].

[1]: https://github.com/cloudfoundry/guardian/blob/c1f268e69cd204e891f29bb020e32284a0054606/gqt/runner/runner.go#L41-L93

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>