[stable/concourse] removes references to fatal errors
Previously, the Helm chart made use of a set of possible fatal errors
that either `garden` or `baggageclaim` would produce, terminating the
worker pod in those cases.

Now, making use of the worker probing endpoint (see [1]), we're able to
implement better strategies for determining whether the worker is up or
not while not changing the contract that `health_ip:health_port` gives
back the info that the worker is alive or not.

[1]: https://github.com/concourse/concourse/commit/c3b26a0cbd9edb533247576d0b2654af1482dc9e

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>